- Updated Examples - Renamed result code wrapper - Introduced disconnect event type

+++ examples/example10.cpp
@@ -92,7 +92,7 @@ int main(int argc, char *argv[]) {
         std::getline(std::cin, input);
 
     } catch (gphoto2pp::exceptions::gphoto2_exception &e) {
-        std::cout << "GPhoto Exception Code: " << e.getResultCode() << std::endl;
+        std::cout << "GPhoto Exception Code: " << static_cast<int>(e.getResultCode()) << std::endl;
         std::cout << "Exception Message: " << e.what() << std::endl;
     }
 }
