+++ examples/example5.cpp
@@ -161,7 +161,7 @@ int main(int argc, char *argv[]) {
             }
         } while (menuNumber != 11);
     } catch (gphoto2pp::exceptions::gphoto2_exception &e) {
-        std::cout << "GPhoto Exception Code: " << e.getResultCode() << std::endl;
+        std::cout << "GPhoto Exception Code: " << static_cast<int>(e.getResultCode()) << std::endl;
         std::cout << "Exception Message: " << e.what() << std::endl;
     }
 }
