+++ examples/example8.cpp
@@ -200,7 +200,7 @@ int main(int argc, char *argv[]) {
         viewFinder.setValue(0);
         cameraWrapper.setConfig(viewFinder);
     } catch (const gphoto2pp::exceptions::gphoto2_exception &e) {
-        std::cout << "GPhoto Exception Code: " << e.getResultCode() << std::endl;
+        std::cout << "GPhoto Exception Code: " << static_cast<int>(e.getResultCode()) << std::endl;
         std::cout << "Exception Message: " << e.what() << std::endl;
     }
 }
